LPD-75060 Created global SelectAccount() on MarketplaceHelper
